requirements.txt
connexion[swagger-ui]

connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
Flask-SQLAlchemy == 2.5.1
SQLAlchemy == 1.4.22
python-jose == 3.1.0
six>=1.9
Flask>=0.10.1
